Add damage control after nervous breakdown
