+++ src/components/Article/ArticleCard.js
@@ -1,0 +1,10 @@
+import React from "react";
+import "./ArticleCard.scss";
+
+const ArticleCard = () => {
+  <div>
+    <p>ArticleCard</p>
+  </div>;
+};
+
+export default ArticleCard;
