+++ src/components/ArticleContainer/ArticleContainer.js
@@ -1,0 +1,13 @@
+import React from "react";
+import ArticleCard from "../Article/ArticleCard";
+import "./ArticleContainer.scss";
+
+const ArticleContainer = () => {
+  return (
+    <div className="card-container">
+      <ArticleCard />
+    </div>
+  );
+};
+
+export default ArticleContainer;
