+++ src/components/ArticleContainer/ArticleContainer.scss
@@ -11,4 +11,7 @@
     width: auto;
     text-align: center;
   }
+  h3{
+    padding: 0;
+  }
 }
