+++ src/components/Routes/Routes.js
@@ -1,16 +1,16 @@
 import React from "react";
 import { Route, Switch } from "react-router-dom";
-import HomeView from "../../Views/HomeView/HomeView";
 import ArticleView from "../../Views/ArticleView/ArticleView";
+import HomeView from "../../Views/HomeView/HomeView";
 
 const Routes = () => {
   return (
     <Switch>
-      <Route exact path="/article/:id">
-        <ArticleView />
-      </Route>
       <Route exact path="/">
         <HomeView />
+      </Route>
+      <Route exact path="/article">
+        <ArticleView />
       </Route>
     </Switch>
   );
